% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatePeaks.R
\name{AnnotatePeaks}
\alias{AnnotatePeaks}
\title{AnnotatePeaks}
\usage{
AnnotatePeaks(peaks_gr, full_gtf_gr)
}
\arguments{
\item{peaks_gr}{A GRanges object representing peak regions.}

\item{full_gtf_gr}{A GRanges object imported from a GTF file containing gene annotations.}
}
\value{
A GRanges object identical to peaks_gr but filtered for having only peaks with a nearest protein-coding gene
}
\description{
AnnotatePeaks
}
