% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_game}
\alias{app_game}
\title{app_game}
\usage{
app_game(input, output, session)
}
\arguments{
\item{input}{UI input}

\item{output}{server output}

\item{session}{number of session}
}
\value{
betting
}
\description{
app_game
}
