% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{Bib}
\alias{Bib}
\title{Consolidate .bib files into a Single File}
\usage{
Bib(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Consolidate .bib files into a Single File
}
\author{
Ivan Jacob Agaloos Pesigan
}
