% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-bib-lib.R
\name{BibLib}
\alias{BibLib}
\title{Copy .bib Files from lib-* GitHUb Repositories}
\usage{
BibLib(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Copy .bib Files from lib-* GitHUb Repositories
}
\author{
Ivan Jacob Agaloos Pesigan
}
