% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{Binary}
\alias{Binary}
\title{Setup Binaries not Included in the Package}
\usage{
Binary(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Setup Binaries not Included in the Package
}
\author{
Ivan Jacob Agaloos Pesigan
}
