% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-build.R
\name{Build}
\alias{Build}
\title{Build Package}
\usage{
Build(path, vignettes = FALSE, dependencies = FALSE, lib_bib = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{vignettes}{Logical.
Build vignettes.}

\item{dependencies}{Logical.
Install package dependencies.}

\item{lib_bib}{Logical.
If \code{lib_bib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Build Package
}
\author{
Ivan Jacob Agaloos Pesigan
}
