% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildignore.R
\name{BuildIgnore}
\alias{BuildIgnore}
\title{Build \code{.Rbuildignore}}
\usage{
BuildIgnore(path, project, add = NULL)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{project}{Character string.
Project name.}

\item{add}{Character vector.
Items to add in \code{.Rbuildignore}.}
}
\description{
Build \code{.Rbuildignore}
}
\author{
Ivan Jacob Agaloos Pesigan
}
