% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-cff.R
\name{CFF}
\alias{CFF}
\title{Build \code{CITATION.cff}}
\usage{
CFF(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Build \code{CITATION.cff}
}
\author{
Ivan Jacob Agaloos Pesigan
}
