% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{Clean}
\alias{Clean}
\title{Clean Files Generated on Build}
\usage{
Clean(path, push = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{push}{Logical.
If \code{push = TRUE} clean in preparation for git push.}
}
\description{
Delete all output files that are generated on automated build.
}
\author{
Ivan Jacob Agaloos Pesigan
}
