% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-pkg.R
\name{CleanPkg}
\alias{CleanPkg}
\title{Clean R Packages}
\usage{
CleanPkg(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Clean R Packages
}
\author{
Ivan Jacob Agaloos Pesigan
}
