% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-config-files.R
\name{ConfigFiles}
\alias{ConfigFiles}
\title{Setup Config Files}
\usage{
ConfigFiles(git_user)
}
\arguments{
\item{git_user}{Character string.
Git user.}
}
\description{
Setup Config Files
}
\details{
This function requires a \code{dotfiles} GitHub repository.
}
\author{
Ivan Jacob Agaloos Pesigan
}
