% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGenomicRanges.R
\name{CreateGenomicRanges}
\alias{CreateGenomicRanges}
\title{CreateGenomicRanges}
\usage{
CreateGenomicRanges(gene_vector, peaks_vector)
}
\arguments{
\item{gene_vector}{A named numeric vector: total expression per each gene.}

\item{peaks_vector}{A named numeric vector: total accessibility per peak region.}
}
\value{
A list containing two GRanges objects:
\describe{
  \item{genes_gr}{GRanges of genes with total expression as metadata}
  \item{peaks_gr}{GRanges of peak regions with total accessibility as metadata}
}
}
\description{
CreateGenomicRanges
}
