% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analysis.R
\name{DataAnalysis}
\alias{DataAnalysis}
\title{Data Analysis}
\usage{
DataAnalysis(path, clean = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{clean}{Logical.
Remove all \code{.Rds} files in \code{data-process} and \code{data-analysis} before running.}
}
\description{
Analyze data in \code{data-process} and store the results as \code{.Rds} files in \code{data-analysis} or \code{.rda} files in \code{data}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
