% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{Dependencies}
\alias{Dependencies}
\title{Copy Dependencies to from \code{.setup/r-dependencies} to \code{R}}
\usage{
Dependencies(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Copy Dependencies to from \code{.setup/r-dependencies} to \code{R}
}
\author{
Ivan Jacob Agaloos Pesigan
}
