% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detritus.R
\name{Detritus}
\alias{Detritus}
\title{Create \verb{detritus/tex/tex-fig-*}}
\usage{
Detritus(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Create \verb{detritus/tex/tex-fig-*}
}
\author{
Ivan Jacob Agaloos Pesigan
}
