% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{download.from.aws.s3}
\alias{download.from.aws.s3}
\title{Download data from Amazon S3 path.}
\usage{
download.from.aws.s3(s3.path, local.path = tempfile(),
  aws.path = "/usr/local/bin")
}
\arguments{
\item{s3.path}{Full path within S3.}

\item{local.path}{Path to location on local drive.}

\item{aws.path}{Path to AWS client.}
}
\value{
Local path to location of downloaded data.
}
\description{
Download data from Amazon S3 path.
}
\examples{
# data.path = download.from.aws.s3("s3://bucket/path/to/data", "~/Desktop/data")
}

