% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawPlot}
\alias{drawPlot}
\title{Draw a scatter plot}
\usage{
drawPlot(df)
}
\arguments{
\item{df}{data.frame with x and y columns}
}
\value{
ggplot object
}
\description{
Draw a scatter plot
}
\examples{
df <- data.frame(x = 1:10, y = 10:1)
drawPlot(df)
}
