% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_plot.R
\name{ds.kcs}
\alias{ds.kcs}
\title{Return a list of the kcs within a partiuclar model.}
\usage{
ds.kcs(stu.step, model)
}
\arguments{
\item{stu.step}{A DataShop Student-Step export data.frame}

\item{model}{The name of a KC model within the dataset}
}
\description{
Return a list of the kcs within a partiuclar model.
}

