% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_plot.R
\name{ds.models}
\alias{ds.models}
\title{Return a list of the available KC models in a Student-Step file.}
\usage{
ds.models(stu.step)
}
\arguments{
\item{stu.step}{A Student-Step export from datashop}
}
\description{
Return a list of the available KC models in a Student-Step file.
}

