% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_plot.R
\name{ds.plot}
\alias{ds.plot}
\title{A Smooth Plot DataShop Learning Curves}
\usage{
ds.plot(stu.step, model, kc = "all", title = NULL, axis.scale = 1)
}
\arguments{
\item{stu.step}{A data.frame of a direct load of a student-step rollup file.}

\item{model}{The name of a KC model to use as it appears in DataShop}

\item{kc}{The name of a particular Kc to get individual kc curves or its number in the list of kcs}

\item{title}{A title for the plot}

\item{axis.scale}{A scaling factor for the y axis of the plot.}
}
\description{
Creates a loess smoothed ggplot from a student step roll up output from Datashop.
}
\keyword{ds.plot}

