% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-project.R
\name{EnvironmentProject}
\alias{EnvironmentProject}
\title{Set \code{PROJECT} Environment Variable}
\usage{
EnvironmentProject(project)
}
\arguments{
\item{project}{Character string.
Project name.}
}
\value{
Returns project name.
}
\description{
Set \code{PROJECT} Environment Variable
}
\author{
Ivan Jacob Agaloos Pesigan
}
