% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-secrets.R
\name{EnvironmentSecrets}
\alias{EnvironmentSecrets}
\title{Set Secrets}
\usage{
EnvironmentSecrets(secrets)
}
\arguments{
\item{secrets}{Named vector of tokens.}
}
\value{
Returns vector of secret tokens.
}
\description{
Set Secrets
}
\author{
Ivan Jacob Agaloos Pesigan
}
