% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-tokens.R
\name{EnvironmentTokens}
\alias{EnvironmentTokens}
\title{Set Tokens}
\usage{
EnvironmentTokens(tokens)
}
\arguments{
\item{tokens}{Named vector of tokens.}
}
\value{
Returns vector of secret tokens.
}
\description{
Set Tokens
}
\author{
Ivan Jacob Agaloos Pesigan
}
