% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{ExtractData}
\alias{ExtractData}
\title{ExtractData}
\usage{
ExtractData(combined_dt)
}
\arguments{
\item{combined_dt}{A combined data table containing feature_id, cell_barcodes and values of both peaks and gene expression.}
}
\value{
Two separate data table, one for gene expression data and one for atac seq data.
}
\description{
ExtractData
}
