% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalizeGeneExpressionGRanges.R
\name{FinalizeGeneExpressionGRanges}
\alias{FinalizeGeneExpressionGRanges}
\title{FinalizeGeneExpressionGRanges}
\usage{
FinalizeGeneExpressionGRanges(genes_gr, full_gtf_gr)
}
\arguments{
\item{genes_gr}{A GRanges object containing gene expression data with `gene_id` metadata.}

\item{full_gtf_gr}{A GRanges object imported from a full GTF annotation file.}
}
\value{
A GRanges object subset of protein-coding genes, with an added metadata column `gene_name`.
}
\description{
FinalizeGeneExpressionGRanges
}
