% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_article.R
\name{fonction_article}
\alias{fonction_article}
\title{Average number of article per month}
\usage{
fonction_article(mois)
}
\arguments{
\item{mois}{the month witch will be considered in the function}
}
\value{
it will return the value of the mean number of article published on LeMonde.fr for the given month
}
\description{
Average number of article per month
}
\examples{
fonction_article(12)
}
