% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_cote.R
\name{fonction_cote}
\alias{fonction_cote}
\title{fonction_cote function}
\usage{
fonction_cote(gareA, gareB, month)
}
\arguments{
\item{gareA}{the name of the departure railway station.
Must be one of the stations in the dataframe \code{SNCF_regularite}.}

\item{gareB}{the name of the arrival railway station.
Must be one of the stations in the dataframe \code{SNCF_regularite}.}

\item{month}{month when the train goes from (\code{gareA} to \code{gareB})}
}
\value{
odd of the train to be late
}
\description{
fonction_cote function
}
\examples{
fonction_cote("PARIS MONTPARNASSE", "NANTES", 12)


}
