% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_raison.R
\name{fonction_raison}
\alias{fonction_raison}
\title{fonction_raison function}
\usage{
fonction_raison(gareA, gareB, month)
}
\arguments{
\item{gareA}{the name of the departure railway station.
Must be one of the stations in the dataframe \code{SNCF_regularite}.}

\item{gareB}{the name of the arrival railway station.
Must be one of the stations in the dataframe \code{SNCF_regularite}.}

\item{month}{month when the train goes from \code{gareA} to \code{gareB}.}
}
\value{
a summary of the most probable reasons for a train to be late
}
\description{
fonction_raison function
}
\examples{
fonction_raison("PARIS MONTPARNASSE", "NANTES", 4)
}
