\name{ftps.to.kmph}
\alias{ftps.to.kmph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Feet per second to kilometers per hour 
}
\description{
The function takes in wind speed value in units, feet per second (ftps) and returns the value after conversion in units, kilometers per hour (kmph)
}
\usage{
ftps.to.kmph(wind.ftps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.ftps}{
  vector of wind speeds in feet per second
}
  \item{round}{
  number of digits after decimal point to be rounded up to
}
}

\value{

\item{wind.kmph }{vector of wind speeds in kilometers per hour}

}

