\name{ftps.to.kts}
\alias{ftps.to.kts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Feet per second to knots
}
\description{
The function takes in wind speed value in units, feet per second (ftps) and returns the value after conversion in units, knots (kts)
}
\usage{
ftps.to.kts(wind.ftps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.ftps}{
vector of wind speeds in feet per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{

 \item{wind.kts }{vector of wind speeds in knots}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

