\name{ftps.to.mph}
\alias{ftps.to.mph}
%% feet per second to miles per hour
\title{
Feet per second to miles per hour
}
\description{
The function takes in wind speed value in units, feet per second (ftps or ft/s) and returns the value after conversion in units, miles per hour (mph)
}
\usage{
ftps.to.mph(wind.ftps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.ftps}{
vector of wind speeds in feet per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.mph }{vector of wind speeds in miles per hour}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
