\name{ftps.to.mps}
\alias{ftps.to.mps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Feet per second to meters per second 
}
\description{
The function takes in wind speed value in units, feet per second (ftps or ft/s) and returns the value after conversion in units, meters per second (mps or m/s)
}
\usage{
ftps.to.mps(wind.ftps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.ftps}{
vector of wind speeds in feet per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.mps }{vector of wind speeds in meters per second}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
