% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateScatterPlot.R
\name{GenerateScatterPlots}
\alias{GenerateScatterPlots}
\title{GenerateScatterPlots}
\usage{
GenerateScatterPlots(plot_data)
}
\arguments{
\item{plot_data}{A data.table or data.frame containing columns: "avg_gex_signal", "avg_atac_signal", "chromosome".}
}
\value{
A list with two ggplot objects:
 \describe{
   \item{total_plot}{Overall scatter plot of gene expression vs. accessibility.}
   \item{faceted_plot}{Scatter plot faceted by chromosome.}
}
}
\description{
GenerateScatterPlots
}
