% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSummaryPlot.R
\name{GenerateSummaryPlots}
\alias{GenerateSummaryPlots}
\title{GenerateSummaryPlots}
\usage{
GenerateSummaryPlots(
  annotated_peaks_gr,
  finalized_expression_gr,
  unassociated_gene_ids
)
}
\arguments{
\item{annotated_peaks_gr}{A GenomicRanges object containing annotated ATAC-seq peaks,with metadata column "total_accessibility" representing peak intensities.}

\item{finalized_expression_gr}{A GenomicRanges object containing gene expression data, with metadata column "total_expression".}

\item{unassociated_gene_ids}{A character vector of gene IDs that are not associated with any ATAC peak.}
}
\value{
A named list containing two ggplot2 plot objects:
\item{peak_intensity_plot}{Boxplot of peak intensities per chromosome.}
\item{unassociated_gex_plot}{Boxplot of gene expression for genes without associated peaks, per chromosome.}
}
\description{
GenerateSummaryPlots
}
