% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{get.all.timestamp.lags}
\alias{get.all.timestamp.lags}
\title{Get timestamp lags for a Sensus data frame.}
\usage{
get.all.timestamp.lags(data)
}
\arguments{
\item{data}{Data to plot lags for (e.g., the result of \code{read.sensus.json}).}
}
\value{
List of lags organized by datum type.
}
\description{
Get timestamp lags for a Sensus data frame.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
lags = get.all.timestamp.lags(data)
plot(lags[["AccelerometerDatum"]])
}

