\name{getMax}
\alias{getMax}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getMax()
}
\description{
function computes the maximun value of a given column in a given dataset
}
\usage{
getMax(datad, columnName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datad}{
 the given dataset
}
\item{columnName}{
the column name from which you want to get the maximum value.
}
}

