% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateAndSummarize.R
\name{IntegrateAndSummarize}
\alias{IntegrateAndSummarize}
\title{IntegrateAndSummarize}
\usage{
IntegrateAndSummarize(
  normalized_gex_dt,
  normalized_atac_dt,
  annotated_peaks_gr_final,
  finalized_expression_gr
)
}
\arguments{
\item{normalized_gex_dt}{A data.table containing normalized gene expression data.It must include a "feature_id" column with gene IDs.}

\item{normalized_atac_dt}{A data.table containing normalized ATAC-seq peak accessibility data.It must include a "feature_id" column with peak IDs.}

\item{annotated_peaks_gr_final}{A GRanges object containing peaks asscoiated with a gene. It must include a "nearest_gene_id" column specifying the closest gene each peak has associated with.}

\item{finalized_expression_gr}{A GRanges object of protein-coding gene expression data including gene IDs in the metadata gene_id column.}
}
\value{
A named list with the following elements:
\describe{
  \item{merged_data}{A data.table merging normalized gene expression and aggregated peak accessibility by gene ID.}
  \item{unmerged_peaks_summary}{A data.table summarizing total, merged, and unmerged peaks counts.}
  \item{unassociated_genes_summary}{A data.table summarizing total genes, genes with associated peaks, and genes without associated peaks.}
  \item{unassociated_gene_ids}{A vector of protein-coding gene IDs (from finalized_expression_gr) with no associated peaks.}
}
}
\description{
IntegrateAndSummarize
}
