\name{kmph.to.kts}
\alias{kmph.to.kts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Kilometers per hour to knots
}
\description{
The function takes in wind speed value in units, kilometers per hour (kmph) and returns the value after conversion in units, knots (kts)
}
\usage{
kmph.to.kts(wind.kmph, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.kmph}{
vector of wind speeds in kilometers per hour
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.kts }{vector of wind speeds in knots}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
