\name{kts.to.ftps}
\alias{kts.to.ftps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Knots to feet per second 
}
\description{
The function takes in wind speed value in units, knots (kts) and returns the value after conversion in units, feet per second (ftps)
}
\usage{
kts.to.ftps(wind.kts, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.kts}{
vector of wind speeds in knots
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.ftps}{vector of wind speeds in feet per second}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
