\name{kts.to.kmph}
\alias{kts.to.kmph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Knots to kilometers per hour
}
\description{
The function takes in wind speed value in units, knots (kts) and returns the value after conversion in units, kilometers per hour (kmph)
}
\usage{
kts.to.kmph(wind.kts, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.kts}{
vector of wind speeds in knots
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.kmph}{vector of wind speeds in kilometers per hour}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
