\name{kts.to.mph}
\alias{kts.to.mph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Knots to miles per hour
}
\description{
The function takes in wind speed value in units, knots (kts) and returns the value after conversion in units, miles per hour (mph)
}
\usage{
kts.to.mph(wind.kts, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.kts}{
vector of wind speeds in knots
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.mph }{vector of wind speeds in miles per hour}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
