\name{kts.to.mps}
\alias{kts.to.mps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Knots to meters per second
}
\description{
The function takes in wind speed value in units, knots (kts) and returns the value after conversion in units, meters per second (mps)
}
\usage{
kts.to.mps(wind.kts, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.kts}{
vector of wind speeds in knots
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.mps}{vector of wind speeds in meters per second}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
