% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-latex-tikz-compile.R
\name{LatexTikzCompile}
\alias{LatexTikzCompile}
\title{Compile Tikz}
\usage{
LatexTikzCompile(path, clean = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{clean}{Logical.
Clean \code{latexmk} output except \verb{*.pdf}}
}
\description{
Compile Tikz
}
\author{
Ivan Jacob Agaloos Pesigan
}
