% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemonde_dates.fr.R
\docType{data}
\name{lemonde_dates}
\alias{lemonde_dates}
\title{Information about the article of Le Monde.}
\format{A data frame with 5032 rows and 34 variables:
\describe{
  \item{month}{month of articles (string, in French)}
  \item{year}{year of the articles}
  \item{n}{number of occurrences}
}}
\source{
\url{https://www.lemonde.fr/}
}
\usage{
lemonde_dates
}
\description{
A dataset containing the information about articles of Le Monde containing words "SNCF" and "grève".
}
\keyword{datasets}
