% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-license.R
\name{License}
\alias{License}
\title{Generate License File}
\usage{
License(path, type = "mit")
}
\arguments{
\item{path}{Character string.
Project path.}

\item{type}{Character string.
License type.
Available values include \code{"mit"} and \code{"gpl3"}.}
}
\description{
Generate License File
}
\author{
Ivan Jacob Agaloos Pesigan
}
