% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{Manual}
\alias{Manual}
\title{Package Manual}
\usage{
Manual(path, project)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{project}{Character string.
Project name.}
}
\description{
Package Manual
}
