\name{mph.to.ftps}
\alias{mph.to.ftps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Miles per hour to feet per second 
}
\description{
The function takes in wind speed value in units, miles per hour (mph) and returns the value after conversion in units, feet per second (ftps or ft/s)
}
\usage{
mph.to.ftps(wind.mph, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mph}{
vector of wind speeds in miles per hour
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.ftps}{vector of wind speeds in feet per second}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
