\name{mph.to.kts}
\alias{mph.to.kts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Miles per hour to knots
}
\description{
The function takes in wind speed value in units, miles per hour (mph) and returns the value after conversion in units, knots (kts)
}
\usage{
mph.to.kts(wind.mph, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mph}{
vector of wind speeds in miles per hour
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.kts}{vector of wind speeds in knots}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
