\name{mps.to.ftps}
\alias{mps.to.ftps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Meters per second to feet per second 
}
\description{
The function takes in wind speed value in units, meters per second (mps or m/s) and returns the value after conversion in units, feet per second (ftps or ft/s)
}
\usage{
mps.to.ftps(wind.mps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mps}{
vector of wind speeds in meters per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.ftps}{vector of wind speeds in feet per second}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
