\name{mps.to.kmph}
\alias{mps.to.kmph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Meters per second to kilometers per hour
}
\description{
The function takes in wind speed value in units, meters per second (mps) and returns the value after conversion in units, kilometers per hour (kmph)
}
\usage{
mps.to.kmph(wind.mps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mps}{
vector of wind speeds in meters per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.kmph}{vector of wind speeds in kilometers per hour}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
