\name{mps.to.kts}
\alias{mps.to.kts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Meters per second to knots
}
\description{
The function takes in wind speed value in units, meters per second (mps)knots (kts) and returns the value after conversion in units, knots (kts)
}
\usage{
mps.to.kts(wind.mps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mps}{
vector of wind speeds in meters per second

}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{wind.kts}{vector of wind speeds in knots}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
