\name{mps.to.mph}
\alias{mps.to.mph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Meters per second to miles per hour
}
\description{
The function takes in wind speed value in units, meters per second (mps) and returns the value after conversion in units, miles per hour (mph)
}
\usage{
mps.to.mph(wind.mps, round = 2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wind.mps}{
vector of wind speeds in meters per second
}
  \item{round}{
number of digits after decimal point to be rounded up to
}
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
 \item{wind.mph}{vector of wind speeds in miles per hour}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
