% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-devel.R
\name{PkgDevel}
\alias{PkgDevel}
\title{Install Development Packages from CRAN}
\usage{
PkgDevel(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Install Development Packages from CRAN
}
\author{
Ivan Jacob Agaloos Pesigan
}
