% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-devel-github.R
\name{PkgDevelGitHub}
\alias{PkgDevelGitHub}
\title{Install Development Packages from GitHub}
\usage{
PkgDevelGitHub(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Install Development Packages from GitHub
}
\author{
Ivan Jacob Agaloos Pesigan
}
