% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-project.R
\name{PkgProject}
\alias{PkgProject}
\title{Install Project Packages from CRAN}
\usage{
PkgProject(path, pkg)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{pkg}{Character vector.
R packages.}
}
\description{
Install Project Packages from CRAN
}
\author{
Ivan Jacob Agaloos Pesigan
}
