% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-pkg-project-bioconductor.R
\name{PkgProjectBioconductor}
\alias{PkgProjectBioconductor}
\title{Install Project Packages from Bioconductor}
\usage{
PkgProjectBioconductor(path, pkg)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{pkg}{Character vector.
Bioconductor packages.}
}
\description{
Install Project Packages from Bioconductor
}
\author{
Ivan Jacob Agaloos Pesigan
}
