% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-pkg-project-github.R
\name{PkgProjectGitHub}
\alias{PkgProjectGitHub}
\title{Install Project Packages from GitHub}
\usage{
PkgProjectGitHub(path, pkg, ref = "HEAD")
}
\arguments{
\item{path}{Character string.
Project path.}

\item{pkg}{Character vector.
GitHub packages.}

\item{ref}{Character vector.
Git branch.}
}
\description{
Install Project Packages from GitHub
}
\author{
Ivan Jacob Agaloos Pesigan
}
