% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-project-version.R
\name{PkgProjectVersion}
\alias{PkgProjectVersion}
\title{Install Project Packages with Specific Version}
\usage{
PkgProjectVersion(path, pkg, ver)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{pkg}{Character vector.
R packages.}

\item{ver}{Character vector.
Version.}
}
\description{
Install Project Packages with Specific Version
}
\author{
Ivan Jacob Agaloos Pesigan
}
