% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.BatteryDatum}
\alias{plot.BatteryDatum}
\title{Plot battery data.}
\usage{
\method{plot}{BatteryDatum}(x, pch = ".", type = "l",
  main = "Battery", ...)
}
\arguments{
\item{x}{Battery data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{main}{Main title.}

\item{...}{Other plotting parameters.}
}
\value{
None
}
\description{
Plot battery data.
}
