% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.LightDatum}
\alias{plot.LightDatum}
\title{Plot light data.}
\usage{
\method{plot}{LightDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Light data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\description{
Plot light data.
}
\examples{
data.path = system.file("extdata", "example_data", package="SensusR")
data = sensus.read.json(data.path)
plot(data$LightDatum)
}
