% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.LocationDatum}
\alias{plot.LocationDatum}
\title{Plot location data.}
\usage{
\method{plot}{LocationDatum}(x, ...)
}
\arguments{
\item{x}{Location data.}

\item{...}{Other plotting parameters.}
}
\description{
Plot location data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$LocationDatum)
}

