% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.ScreenDatum}
\alias{plot.ScreenDatum}
\title{Plot screen data.}
\usage{
\method{plot}{ScreenDatum}(x, ...)
}
\arguments{
\item{x}{Screen data.}

\item{...}{Other plotting parameters.}
}
\description{
Plot screen data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$ScreenDatum)
}

