% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.SpeedDatum}
\alias{plot.SpeedDatum}
\title{Plot speed data.}
\usage{
\method{plot}{SpeedDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Speed data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\description{
Plot speed data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$SpeedDatum)
}

