% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.WlanDatum}
\alias{plot.WlanDatum}
\title{Plot WLAN data.}
\usage{
\method{plot}{WlanDatum}(x, ...)
}
\arguments{
\item{x}{WLAN data.}

\item{...}{Other plotting parameters.}
}
\value{
None
}
\description{
Plot WLAN data.
}
