% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{Profile}
\alias{Profile}
\title{Setup \code{.Rprofile}}
\usage{
Profile(path, project)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{project}{Character string.
Project name.}
}
\description{
Setup \code{.Rprofile}
}
\author{
Ivan Jacob Agaloos Pesigan
}
