% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{Setup the project.Rproj file}
\usage{
Project(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Setup the project.Rproj file
}
\author{
Ivan Jacob Agaloos Pesigan
}
