% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_findfile}
\alias{project_findfile}
\title{Search the file from current folder to root}
\usage{
project_findfile(filename, base = getwd())
}
\arguments{
\item{filename}{the filename for search}
}
\description{
Search the file from current folder to root
}

