% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_fun}
\alias{project_fun}
\title{Source the function in the project}
\usage{
project_fun(force = FALSE, project = NULL, all = FALSE)
}
\arguments{
\item{force}{Force to reload all functions}

\item{project}{The project name}

\item{all}{Source all functions}
}
\description{
Source the function in the project
}

