% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_get_para}
\alias{project_get_para}
\title{Get parameter from config file}
\usage{
project_get_para(name = NULL, project = NULL)
}
\arguments{
\item{name}{The parameter name}

\item{project}{The project name}
}
\description{
Get parameter from config file
}

