% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_read}
\alias{project_read}
\title{Read a project file (csv, RData, ...)}
\usage{
project_read(file, project = NULL, var = NULL)
}
\arguments{
\item{file}{a character string giving the name of the file to load}

\item{project}{The project name}
}
\description{
Read a project file (csv, RData, ...)
}

