% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_set_pata}
\alias{project_set_pata}
\title{Set parameter}
\usage{
project_set_pata(name, value, project = NULL)
}
\arguments{
\item{name}{The parameter name}

\item{value}{The parameter value}

\item{project}{The project name}
}
\description{
Set parameter
}

