% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_variable_name}
\alias{project_variable_name}
\title{Generate variable name for global environment}
\usage{
project_variable_name(project = NULL)
}
\arguments{
\item{project}{The project name}
}
\description{
Generate variable name for global environment
}
