% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-quarto.R
\name{Quarto}
\alias{Quarto}
\title{Render Quarto Project}
\usage{
Quarto(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Render Quarto Project
}
\author{
Ivan Jacob Agaloos Pesigan
}
