% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{read.sensus.json}
\alias{read.sensus.json}
\title{Read JSON-formatted Sensus data.}
\usage{
read.sensus.json(data.path, is.directory = TRUE, recursive = TRUE,
  convert.to.local.timezone = TRUE, local.timezone = Sys.timezone())
}
\arguments{
\item{data.path}{Path to Sensus JSON data (either a file or a directory).}

\item{is.directory}{Whether or not the path is a directory.}

\item{recursive}{Whether or not to read files recursively from directory indicated by path.}

\item{convert.to.local.timezone}{Whether or not to convert timestamps to the local timezone.}

\item{local.timesonze}{If converting timestamps to local timesonze, the local timezone to use.}
}
\value{
All data, listed by type.
}
\description{
Read JSON-formatted Sensus data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"), is.directory = FALSE)
}

