% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-readme.R
\name{ReadMe}
\alias{ReadMe}
\title{Build README.md}
\usage{
ReadMe(path, bib_lib = TRUE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{bib_lib}{Logical.
If \code{bib_lib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Build README.md
}
\author{
Ivan Jacob Agaloos Pesigan
}
