% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{Repository}
\alias{Repository}
\title{Setup RStudio Package Manager Repository Based on the Current Operating System}
\usage{
Repository(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\value{
Returns the repository url.
}
\description{
Setup RStudio Package Manager Repository Based on the Current Operating System
}
\author{
Ivan Jacob Agaloos Pesigan
}
