% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocoding.R
\name{rev_geocoding}
\alias{rev_geocoding}
\title{rev_geocoding}
\usage{
rev_geocoding(lat, lon)
}
\arguments{
\item{lat}{latitude of a place}

\item{lon}{longitude of a place}
}
\value{
postcode of input address
}
\description{
rev_geocoding
}
\examples{
rev_geocoding(47.01671,-1.435332)
}
