% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocoding_country.R
\name{rev_geocoding_country}
\alias{rev_geocoding_country}
\title{rev_geocoding_country}
\usage{
rev_geocoding_country(lat, lon)
}
\arguments{
\item{lat}{latitude of the departement}

\item{lon}{longitude of the departement}
}
\value{
country (char)
}
\description{
rev_geocoding_country
}
\examples{
rev_geocoding_country( 47.01671,-1.435332)

}
