% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.decompress.json}
\alias{sensus.decompress.json}
\title{Decompresses JSON files downloaded from AWS S3.}
\usage{
sensus.decompress.json(local.path)
}
\arguments{
\item{local.path}{Path to location on local machine.}
}
\value{
None
}
\description{
Decompresses JSON files downloaded from AWS S3.
}
\examples{
# sensus.decompress("~/Desktop/data")
}

