% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.get.all.timestamp.lags}
\alias{sensus.get.all.timestamp.lags}
\title{Get timestamp lags for a Sensus data frame.}
\usage{
sensus.get.all.timestamp.lags(data)
}
\arguments{
\item{data}{Data to plot lags for (e.g., the result of \code{sensus.read.json.files}).}
}
\value{
List of lags organized by datum type.
}
\description{
Get timestamp lags for a Sensus data frame.
}
