% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.get.timestamp.lags}
\alias{sensus.get.timestamp.lags}
\title{Get timestamp lags for a Sensus datum.}
\usage{
sensus.get.timestamp.lags(datum)
}
\arguments{
\item{datum}{One element of a Sensus data frame (e.g., data$CompassDatum).}
}
\value{
List of lags.
}
\description{
Get timestamp lags for a Sensus datum.
}
\examples{
data.path = system.file("extdata", "example-data", package="SensusR")
data = sensus.read.json.files(data.path)
plot(sensus.get.timestamp.lags(data$AccelerometerDatum))
}
