% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.get.unique.device.ids}
\alias{sensus.get.unique.device.ids}
\title{Gets unique device IDs within a dataset.}
\usage{
sensus.get.unique.device.ids(data)
}
\arguments{
\item{data}{Data to write, as read using \code{\link{sensus.read.json.files}}.}
}
\value{
Unique device IDs within the data.
}
\description{
Gets unique device IDs within a dataset.
}
