% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.list.aws.s3.buckets}
\alias{sensus.list.aws.s3.buckets}
\title{Lists S3 buckets.}
\usage{
sensus.list.aws.s3.buckets(profile = "default",
  aws.path = "/usr/local/bin/aws")
}
\arguments{
\item{profile}{AWS credentials profile to use for authentication.}

\item{aws.path}{Path to AWS client.}
}
\value{
None
}
\description{
Lists S3 buckets.
}
