% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.plot.data.frequency.by.day}
\alias{sensus.plot.data.frequency.by.day}
\title{Plot data frequency by day.}
\usage{
sensus.plot.data.frequency.by.day(datum, xlab = "Study Day",
  ylab = "Data Frequency", main = "Data Frequency")
}
\arguments{
\item{datum}{Data frame for a single datum.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{main}{Label for plot.}
}
\description{
Plot data frequency by day.
}
\examples{
data.path = system.file("extdata", "example_data", package="SensusR")
data = sensus.read.json.files(data.path)
sensus.plot.data.frequency.by.day(data$AccelerometerDatum)
}
