% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.r
\name{source_dir}
\alias{source_dir}
\title{Sourcing a bunch of files located from different folders/subfolders.}
\usage{
source_dir(path, recursive = TRUE, only.funs = TRUE, ...)
}
\arguments{
\item{path}{"Character". It is the path you want to source files.}

\item{recursive}{"Logical". If TRUE, files under subfolders will also be sourced.}

\item{only.funs}{Locigal. If TRUE, only functions will be sourced.}

\item{...}{type help("source")}
}
\description{
Sourcing a bunch of files located from different folders/subfolders.
}
