% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sourceDir.r
\name{source_funs}
\alias{source_funs}
\title{Source functions from a File or a connection}
\usage{
source_funs(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), prompt.echo = getOption("prompt"),
  max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), continue.echo = getOption("continue"),
  skip.echo = 0, keep.source = getOption("keep.source"))
}
\arguments{
\item{file}{a connection or a character string giving the pathname of the file or URL to read from. "" indicates the connection stdin().}

\item{local}{See source}

\item{echo}{See source}

\item{print.eval}{See source}

\item{verbose}{See source}

\item{prompt.echo}{See source}

\item{max.deparse.length}{See source}

\item{chdir}{See source}

\item{encoding}{See source}

\item{continue.echo}{See source}

\item{skip.echo}{See source}

\item{keep.source}{See source}
}
\description{
This function is same as source, only including functions. Modified from source function in R
}

