% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-tinytex.R
\name{TinyTex}
\alias{TinyTex}
\title{Install \verb{~/.TinyTex}}
\usage{
TinyTex(path, bundle = "TinyTeX-2", force = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{bundle}{The bundle name of TinyTeX (which determines the collection of
LaTeX packages to install). See
\url{https://github.com/rstudio/tinytex-releases#releases} for all possible
bundles and their meanings.}

\item{force}{Logical.
Force install.}
}
\description{
Install \verb{~/.TinyTex}
}
\author{
Ivan Jacob Agaloos Pesigan
}
