% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalAccessibilityPerPeak.R
\name{TotalAccessibilityPerPeak}
\alias{TotalAccessibilityPerPeak}
\title{TotalAccessibilityPerPeak}
\usage{
TotalAccessibilityPerPeak(atac_seq_dt)
}
\arguments{
\item{atac_seq_dt}{A data.table with peak regions as rows and cells as columns. The first column should be 'feature_id'.}
}
\value{
A named numeric vector: total accessibility per peak region.
}
\description{
TotalAccessibilityPerPeak
}
