% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalExpressionPerGene.R
\name{TotalExpressionPerGene}
\alias{TotalExpressionPerGene}
\title{TotalExpressionPerGene}
\usage{
TotalExpressionPerGene(gene_expression_dt)
}
\arguments{
\item{gene_expression_dt}{A data.table with gene ids (and gene expression) as rows and cells as columns. The first column should be 'feature_id'. Previously assigned from ExtractData as gene_expression_dt <- results$gene_expression_dt .}
}
\value{
A named numeric vector: total accessibility per peak region.
}
\description{
TotalExpressionPerGene
}
