% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{trim.leading}
\alias{trim.leading}
\title{Trim leading white space from a string.}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{String to trim.}
}
\value{
Result of trimming.
}
\description{
Trim leading white space from a string.
}
