% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing white space from a string.}
\usage{
trim(x)
}
\arguments{
\item{x}{String to trim.}
}
\value{
Result of trimming.
}
\description{
Trim leading and trailing white space from a string.
}
\examples{
trim("  asdf  ")
}
