% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{trim.trailing}
\alias{trim.trailing}
\title{Trim trailing white space from a string.}
\usage{
trim.trailing(x)
}
\arguments{
\item{x}{String to trim.}
}
\value{
Result of trimming.
}
\description{
Trim trailing white space from a string.
}

