% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes-precompile.R
\name{VignettesPrecompile}
\alias{VignettesPrecompile}
\title{Precompile Vignettes}
\usage{
VignettesPrecompile(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Precompile Vignettes
}
\author{
Ivan Jacob Agaloos Pesigan
}
