% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-bib.R
\name{Bib}
\alias{Bib}
\title{Consolidate .bib files into a Single File}
\usage{
Bib(path, bib_lib = TRUE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{bib_lib}{Logical.
If \code{bib_lib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Consolidate .bib files into a Single File
}
\author{
Ivan Jacob Agaloos Pesigan
}
