% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{Build}
\alias{Build}
\title{Build Package}
\usage{
Build(path, vignettes = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{vignettes}{Logical.
Build vignettes.}
}
\description{
Build Package
}
\author{
Ivan Jacob Agaloos Pesigan
}
