% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-clean-proj.R
\name{CleanProj}
\alias{CleanProj}
\title{Clean Project}
\usage{
CleanProj(path, lib = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{lib}{Logical.
If \code{lib = TRUE}, delete \code{.library}.
If \code{lib = FALSE}, \code{.library} is not deleted.}
}
\description{
Remove all files except \code{project.Rproj} and,
optionally, \code{.library}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
