% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-tinytex.R
\name{CleanTinyTex}
\alias{CleanTinyTex}
\title{Clean \verb{~/.TinyTex}}
\usage{
CleanTinyTex(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Uninstall \code{TinyTex}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
