% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{DataProcess}
\alias{DataProcess}
\title{Process Data}
\usage{
DataProcess(path, clean = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{clean}{Logical.
Remove all \code{.Rds} files in \code{data-process} before running.}
}
\description{
Execute R scripts in \code{data-process}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
