% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{Environment}
\alias{Environment}
\title{Setup \code{.Renviron}}
\usage{
Environment(path, project, secrets)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{project}{Character string.
Project name.}

\item{secrets}{Named vector of tokens.}
}
\description{
Setup \code{.Renviron}
}
\author{
Ivan Jacob Agaloos Pesigan
}
