% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData.R
\name{ExtractData}
\alias{ExtractData}
\title{ExtractData}
\usage{
ExtractData(combined_dt)
}
\arguments{
\item{combined_dt}{A combined data table containing feature_id, cell_barcodes and values of both peaks and gene expression.}
}
\value{
A list with two data.tables:
\describe{
  \item{gene_expression_dt}{Contains rows with `feature_id` starting with "ENSG", representing gene expression data}
  \item{atac_seq_dt}{Contains rows with `feature_id` starting with "chr", representing chromatin accessibility data}
}
}
\description{
ExtractData
}
