% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-knit.R
\name{LatexKnit}
\alias{LatexKnit}
\title{Knit \verb{latexsrc/*.Rtex}}
\usage{
LatexKnit(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Knit \verb{latexsrc/*.Rtex}
}
\author{
Ivan Jacob Agaloos Pesigan
}
