% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-latex-make.R
\name{LatexMake}
\alias{LatexMake}
\title{Compile Output of \code{LatexKnit}}
\usage{
LatexMake(path, clean = FALSE, bib_lib = TRUE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{clean}{Logical.
Clean \code{latexmk} output except \verb{*.pdf}}

\item{bib_lib}{Logical.
If \code{bib_lib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Compile Output of \code{LatexKnit}
}
\author{
Ivan Jacob Agaloos Pesigan
}
