% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-make.R
\name{LatexMake}
\alias{LatexMake}
\title{Compile Output of \code{LatexKnit}}
\usage{
LatexMake(path, clean = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{clean}{Logical.
Clean \code{latexmk} output except \verb{*.pdf}}
}
\description{
Compile Output of \code{LatexKnit}
}
\author{
Ivan Jacob Agaloos Pesigan
}
