% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-project-github.R
\name{PkgProjectGitHub}
\alias{PkgProjectGitHub}
\title{Install Project Packages from GitHub}
\usage{
PkgProjectGitHub(path, pkg)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{pkg}{Character vector.
GitHub packages.}
}
\description{
Install Project Packages from GitHub
}
\author{
Ivan Jacob Agaloos Pesigan
}
