% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.BatteryDatum}
\alias{plot.BatteryDatum}
\title{Plot battery data.}
\usage{
\method{plot}{BatteryDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Battery data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\description{
Plot battery data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$BatteryDatum)
}

