% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.CellTowerDatum}
\alias{plot.CellTowerDatum}
\title{Plot cell tower data.}
\usage{
\method{plot}{CellTowerDatum}(x, ...)
}
\arguments{
\item{x}{Cell tower data.}

\item{...}{Other plotting arguments.}
}
\description{
Plot cell tower data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$CellTowerDatum)
}

