% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.CompassDatum}
\alias{plot.CompassDatum}
\title{Plot compass data.}
\usage{
\method{plot}{CompassDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Compass data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\value{
None
}
\description{
Plot compass data.
}
