% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.LightDatum}
\alias{plot.LightDatum}
\title{Plot light data.}
\usage{
\method{plot}{LightDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Light data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\description{
Plot light data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$LightDatum)
}

