% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.LocationDatum}
\alias{plot.LocationDatum}
\title{Plot location data.}
\usage{
\method{plot}{LocationDatum}(x, ...)
}
\arguments{
\item{x}{Location data.}

\item{...}{Arguments to pass to plotting routines. This can include two special arguments:  qmap.args (passed to \code{\link{qmap}}) and geom.point.args (passed to \code{\link{geom_point}}).}
}
\description{
Plot location data.
}
\examples{
data.path = system.file("extdata", "example_data", package="SensusR")
data = sensus.read.json.files(data.path)
plot(data$LocationDatum)
}
