% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.ScreenDatum}
\alias{plot.ScreenDatum}
\title{Plot screen data.}
\usage{
\method{plot}{ScreenDatum}(x, ...)
}
\arguments{
\item{x}{Screen data.}

\item{...}{Other plotting parameters.}
}
\description{
Plot screen data.
}
\examples{
data.path = system.file("extdata", "example_data", package="SensusR")
data = sensus.read.json.files(data.path)
plot(data$ScreenDatum)
}
