% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.SoundDatum}
\alias{plot.SoundDatum}
\title{Plot sound data.}
\usage{
\method{plot}{SoundDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Sound data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\value{
None
}
\description{
Plot sound data.
}
