% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.TelephonyDatum}
\alias{plot.TelephonyDatum}
\title{Plot telephony data.}
\usage{
\method{plot}{TelephonyDatum}(x, ...)
}
\arguments{
\item{x}{Telephony data.}

\item{...}{Other plotting parameters.}
}
\value{
None
}
\description{
Plot telephony data.
}
