% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.WlanDatum}
\alias{plot.WlanDatum}
\title{Plot WLAN data.}
\usage{
\method{plot}{WlanDatum}(x, ...)
}
\arguments{
\item{x}{WLAN data.}

\item{...}{Other plotting parameters.}
}
\description{
Plot WLAN data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$WlanDatum)
}

