% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_filepath}
\alias{project_filepath}
\title{Construct the path to a file from components in the project}
\usage{
project_filepath(..., project = NULL)
}
\arguments{
\item{...}{character vectors}

\item{project}{The project name}
}
\description{
Construct the path to a file from components in the project
}

