% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_findfile}
\alias{project_findfile}
\title{Search the file from current folder to root}
\usage{
project_findfile(filename, base = getwd())
}
\arguments{
\item{filename}{the filename for search}
}
\description{
Search the file from current folder to root
}
