% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project_get_para}
\alias{project_get_para}
\title{Get parameter from config file}
\usage{
project_get_para(name = NULL, project = NULL)
}
\arguments{
\item{name}{The parameter name}

\item{project}{The project name}
}
\description{
Get parameter from config file
}

