% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project_load}
\alias{project_load}
\title{Load a project, the first function will be called when working for a project}
\usage{
project_load(project = NULL, base = getwd())
}
\arguments{
\item{project}{The project name}
}
\description{
Load a project, the first function will be called when working for a project
}

