% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project_read}
\alias{project_read}
\title{Read a project file (csv, RData, ...)}
\usage{
project_read(file, project = NULL, var = NULL)
}
\arguments{
\item{file}{a character string giving the name of the file to load}

\item{project}{The project name}
}
\description{
Read a project file (csv, RData, ...)
}

