% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project_read_para}
\alias{project_read_para}
\title{Read parameter files}
\usage{
project_read_para(project = NULL, base = NULL)
}
\arguments{
\item{project}{The project name}
}
\description{
Read parameter files
}

