% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-quarto.R
\name{Quarto}
\alias{Quarto}
\title{Render Quarto Project}
\usage{
Quarto(path, bib_lib = TRUE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{bib_lib}{Logical.
If \code{bib_lib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Render Quarto Project
}
\author{
Ivan Jacob Agaloos Pesigan
}
