% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-readme.R
\name{ReadMe}
\alias{ReadMe}
\title{Build README.md}
\usage{
ReadMe(path)
}
\arguments{
\item{path}{Character string.
Project path.}
}
\description{
Build README.md
}
\author{
Ivan Jacob Agaloos Pesigan
}
