% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.get.all.timestamp.lags}
\alias{sensus.get.all.timestamp.lags}
\title{Get timestamp lags for a Sensus data frame.}
\usage{
sensus.get.all.timestamp.lags(data)
}
\arguments{
\item{data}{Data to plot lags for (e.g., the result of \code{read.sensus.json}).}
}
\value{
List of lags organized by datum type.
}
\description{
Get timestamp lags for a Sensus data frame.
}
\examples{
data.path = system.file("extdata", "example_data", package="SensusR")
data = sensus.read.json.files(data.path)
lags = sensus.get.all.timestamp.lags(data)
plot(lags[["AccelerometerDatum"]])
}
