% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\name{sensus.get.timestamp.lags}
\alias{sensus.get.timestamp.lags}
\title{Get timestamp lags for a Sensus datum.}
\usage{
sensus.get.timestamp.lags(datum)
}
\arguments{
\item{datum}{Data to plot lags for (e.g., the result of \code{sensus.read.json.files}).}
}
\value{
List of lags.
}
\description{
Get timestamp lags for a Sensus datum.
}
