% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensusR.R
\docType{package}
\name{SensusR}
\alias{SensusR}
\alias{SensusR-package}
\title{SensusR:  Sensus Analytics}
\description{
Provides access and analytic functions for Sensus data. More information can be found at the
following URL:
}
\details{
https://predictive-technology-laboratory.github.io/sensus
}
\section{SensusR functions}{

The SensusR functions handle reading, cleaning, plotting, and otherwise analyzing data collected
via the Sensus system.
}

