% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rProject-site.R
\name{Site}
\alias{Site}
\title{Build Website}
\usage{
Site(path, lib_bib = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{lib_bib}{Logical.
If \code{lib_bib = TRUE},
build \code{bib.bib} from \verb{lib-*} GitHub repositories.}
}
\description{
Build Website
}
\author{
Ivan Jacob Agaloos Pesigan
}
