% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytex.R
\name{TinyTex}
\alias{TinyTex}
\title{Install \verb{~/.TinyTex}}
\usage{
TinyTex(path, force = FALSE)
}
\arguments{
\item{path}{Character string.
Project path.}

\item{force}{Logical.
Force install.}
}
\description{
Install \verb{~/.TinyTex}
}
\author{
Ivan Jacob Agaloos Pesigan
}
